refactor(tesseratos): make it a lib, move tools
