+++ engine/CMakeLists.txt
@@ -22,15 +22,6 @@ set(CUBOS_ENGINE_SOURCE
     "src/cubos/engine/imgui/imgui.cpp"
     "src/cubos/engine/imgui/serialization.cpp"
 
-    "src/cubos/engine/tools/asset_explorer/plugin.cpp"
-    "src/cubos/engine/tools/settings_inspector/plugin.cpp"
-    "src/cubos/engine/tools/entity_selector/plugin.cpp"
-    "src/cubos/engine/tools/world_inspector/plugin.cpp"
-    "src/cubos/engine/tools/entity_inspector/plugin.cpp"
-    "src/cubos/engine/tools/scene_editor/plugin.cpp"
-    "src/cubos/engine/tools/debug_camera/plugin.cpp"
-    "src/cubos/engine/tools/voxel_palette_editor/plugin.cpp"
-
     "src/cubos/engine/transform/plugin.cpp"
     "src/cubos/engine/transform/local_to_world.cpp"
     "src/cubos/engine/transform/position.cpp"
