--- engine/include/cubos/engine/tools/module.dox
@@ -1,11 +1,0 @@
-/// @dir
-/// @brief @ref tool-plugins plugins directory.
-
-#pragma once
-
-namespace cubos::engine
-{
-    /// @defgroup tool-plugins Tools
-    /// @ingroup engine
-    /// @brief Plugins which add debug and development tooling to @b CUBOS.
-}
