--- engine/src/cubos/engine/tools/entity_selector/plugin.cpp
@@ -1,6 +1,0 @@
-#include <cubos/engine/tools/entity_selector/plugin.hpp>
-
-void cubos::engine::tools::entitySelectorPlugin(Cubos& cubos)
-{
-    cubos.addResource<cubos::engine::tools::EntitySelector>();
-}
