+++ tools/tesseratos/include/tesseratos/asset_explorer/plugin.hpp
@@ -1,0 +1,35 @@
+/// @dir
+/// @brief @ref tesseratos-asset-explorer-plugin plugin directory.
+
+/// @file
+/// @brief Plugin entry point.
+/// @ingroup asset-explorer-plugin
+
+#pragma once
+
+#include <cubos/engine/assets/plugin.hpp>
+
+namespace tesseratos
+{
+    /// @defgroup tesseratos-asset-explorer-plugin Asset explorer
+    /// @ingroup tesseratos
+    /// @brief Allows viewing and selecting assets through a ImGui window.
+    ///
+    /// ## Events
+    /// - @ref AssetSelectedEvent - sent when an asset is selected in the explorer.
+    ///
+    /// ## Dependencies
+    /// - @ref imgui-plugin
+    /// - @ref assets-plugin
+
+    /// @brief Event sent when an asset is selected.
+    struct AssetSelectedEvent
+    {
+        cubos::engine::AnyAsset asset; ///< Handle to the selected asset.
+    };
+
+    /// @brief Plugin entry function.
+    /// @param cubos @b CUBOS. main class
+    /// @ingroup tesseratos-asset-explorer-plugin
+    void assetExplorerPlugin(cubos::engine::Cubos& cubos);
+} // namespace tesseratos
