+++ tools/tesseratos/include/tesseratos/settings_inspector/plugin.hpp
@@ -1,0 +1,26 @@
+/// @dir
+/// @brief @ref tesseratos-settings-inspector-plugin plugin directory.
+
+/// @file
+/// @brief Plugin entry point.
+/// @ingroup tesseratos-settings-inspector-plugin
+
+#pragma once
+
+#include <cubos/engine/cubos.hpp>
+
+namespace tesseratos
+{
+    /// @defgroup tesseratos-settings-inspector-plugin Settings inspector
+    /// @ingroup tesseratos
+    /// @brief Allows inspecting the current setting values through a ImGui window.
+    ///
+    /// ## Dependencies
+    /// - @ref settings-plugin
+    /// - @ref imgui-plugin
+
+    /// @brief Plugin entry function.
+    /// @param cubos @b CUBOS. main class
+    /// @ingroup tesseratos-settings-inspector-plugin
+    void settingsInspectorPlugin(cubos::engine::Cubos& cubos);
+} // namespace tesseratos
