+++ tools/tesseratos/src/tesseratos/entity_selector/plugin.cpp
@@ -1,0 +1,8 @@
+#include <tesseratos/entity_selector/plugin.hpp>
+
+using cubos::engine::Cubos;
+
+void tesseratos::entitySelectorPlugin(Cubos& cubos)
+{
+    cubos.addResource<EntitySelector>();
+}
